+++ project/app/src/test/java/org/owntracks/android/logging/LogRingBufferTest.kt
@@ -7,7 +7,7 @@ import org.junit.Test
 class LogRingBufferTest {
     @Test
     fun `Given an empty buffer, when adding an entry, then the buffer has one entry`() {
-        val logEntry = LogEntry(3, "tag", "message")
+        val logEntry = LogEntry(3, "tag", "message", "main")
         val logRingBuffer = LogRingBuffer(5)
         logRingBuffer.add(logEntry)
         assertEquals(1, logRingBuffer.all().size)
@@ -16,7 +16,7 @@ class LogRingBufferTest {
 
     @Test
     fun `Given a full buffer, when adding a new log entry, then the buffer does not contain the first entry`() {
-        val logEntries = (0..5).map { LogEntry(it, "tag", "message") }
+        val logEntries = (0..5).map { LogEntry(it, "tag", "message", "main") }
         val logRingBuffer = LogRingBuffer(5)
         logEntries.forEach { logRingBuffer.add(it) }
         assertEquals(5, logRingBuffer.all().size)
@@ -25,7 +25,7 @@ class LogRingBufferTest {
 
     @Test
     fun `Given a full buffer, when requesting the logs, they are returned in the correct order`() {
-        val logEntries = (0..12).map { LogEntry(it, "tag", "message") }
+        val logEntries = (0..12).map { LogEntry(it, "tag", "message", "main") }
         val logRingBuffer = LogRingBuffer(5)
         logEntries.forEach { logRingBuffer.add(it) }
         assertEquals(8, logRingBuffer.all().first().priority)
@@ -34,7 +34,7 @@ class LogRingBufferTest {
 
     @Test
     fun `Given a partial buffer, when requesting the logs, they are returned in the correct order`() {
-        val logEntries = (0..9).map { LogEntry(it, "tag", "message") }
+        val logEntries = (0..9).map { LogEntry(it, "tag", "message", "main") }
         val logRingBuffer = LogRingBuffer(20)
         logEntries.forEach { logRingBuffer.add(it) }
         assertEquals(0, logRingBuffer.all().first().priority)
@@ -43,7 +43,7 @@ class LogRingBufferTest {
 
     @Test
     fun `Given a full buffer, when clearing, then the buffer is empty`() {
-        val logEntries = (0..5).map { LogEntry(it, "tag", "message") }
+        val logEntries = (0..5).map { LogEntry(it, "tag", "message", "main") }
         val logRingBuffer = LogRingBuffer(5)
         logEntries.forEach { logRingBuffer.add(it) }
         logRingBuffer.clear()